Split soundboard types
