+++ packages/types/src/index.ts
@@ -15,6 +15,7 @@ export * from './discord/messages.js'
 export * from './discord/polls.js'
 export * from './discord/reference.js'
 export * from './discord/skus.js'
+export * from './discord/soundboards.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
